refactor(training): infrastructure/repository decomposto por agregado (Fase 5.4)
Substitui `src/training/infrastructure/repository.py` (monolítico, 739
linhas, 13 Repository classes) por pacote `infrastructure/repository/`
com 8 submódulos por agregado + `__init__.py` que re-exporta toda a
surface pública.

Submódulos:
- sessions.py       → TrainingSessionRepository + SessionObjectiveRepository
                      (inclui helper privado `_to_decimal`)
- blocks.py         → SessionBlockRepository
- wellness.py       → WellnessPreRepository + WellnessPostRepository
- attendance.py     → AttendanceRepository
- execution.py      → ExecutionRecordRepository
- communication.py  → FeedbackThreadRepository + AttentionQueueRepository
                      + RecommendationRepository
- eligibility.py    → AthleteIneligibilityDeclarationRepository
- planning.py       → MesocycleRepository + MicrocycleRepository

Zero-gap:
- `training.infrastructure.repository.<RepoClass>` continua resolvendo
  (package `__init__` re-exporta as 13 classes). application/, api/ e
  testes não mudam.
- Suite training: 375 passed, 19 skipped (idêntico a Fase 5.3).

Motivação: o arquivo atingiu 29KB/739 linhas com padrão repetitivo
(save/list/get_by_id/_to_domain) por agregado. A decomposição casa com
o layout de `domain/entities/` (Fase 5.3) e `application/` (Fase 3),
uniformizando um eixo de navegação por agregado em todas as camadas.
